+++ lib/data/models/trending_recipe_model.dart
@@ -27,4 +27,5 @@ class TrendingRecipeModel {
       rating: json['rating'],
     );
   }
+
 }
